+++ starknet_py/net/gateway_client.py
@@ -7,7 +7,7 @@ from marshmallow import EXCLUDE
 from starkware.starknet.services.api.gateway.transaction import AccountTransaction
 
 from starknet_py.net.client import Client
-from starknet_py.net.client_models import (
+from starknet_py.net.models.client import (
     Transaction,
     SentTransactionResponse,
     ContractCode,
