+++ starknet_py/net/models/client.py
@@ -12,7 +12,6 @@ from starkware.starknet.services.api.gateway.transaction import (
     Declare as DCL,
 )
 
-
 from starknet_py.utils.docs import as_our_module
 
 InvokeFunction = as_our_module(IF)
@@ -20,7 +19,6 @@ StarknetTransaction = as_our_module(T)
 ContractClass = as_our_module(CD)
 Deploy = as_our_module(D)
 Declare = as_our_module(DCL)
-
 
 Hash = Union[int, str]
 Tag = Literal["pending", "latest"]
