+++ starknet_py/net/schemas/common_test.py
@@ -4,7 +4,7 @@ from marshmallow import Schema, ValidationError
 
 from starknet_py.net.schemas.common import NonPrefixedHex
 from starknet_py.net.schemas.rpc import Felt, StatusField, BlockStatusField
-from starknet_py.net.client_models import TransactionStatus, BlockStatus
+from starknet_py.net.models.client import TransactionStatus, BlockStatus
 
 
 def test_serialize_felt():
