+++ starknet_py/tests/e2e/client/full_node_test.py
@@ -4,7 +4,7 @@ from unittest.mock import patch, MagicMock
 import pytest
 
 from starknet_py.net.client_errors import ClientError
-from starknet_py.net.client_models import (
+from starknet_py.net.models.client import (
     DeployTransaction,
 )
