+++ starknet_py/transactions/declare.py
@@ -7,7 +7,7 @@ from starkware.starknet.services.api.gateway.transaction import (
 
 from starknet_py.common import create_compiled_contract
 from starknet_py.compile.compiler import StarknetCompilationSource
-from starknet_py.net.client_models import Declare
+from starknet_py.net.models.client import Declare
 
 
 def make_declare_tx(
