+++ starknet_py/transactions/deploy.py
@@ -5,7 +5,7 @@ from starkware.starknet.definitions.fields import ContractAddressSalt
 
 from starknet_py.common import create_compiled_contract
 from starknet_py.compile.compiler import StarknetCompilationSource
-from starknet_py.net.client_models import ContractClass
+from starknet_py.net.models.client import ContractClass
 from starknet_py.net.models import Deploy
